05-monitoring/5.2_monitoring_ground_truth/requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
dataset==2.20.0
ragas==0.1.9
elasticsearch==8.14.0